+++ include/mujoco/mjmodel.h
@@ -111,6 +111,7 @@ typedef enum mjtGeom_ {           // type of geometric shape
   mjGEOM_LINE,                    // line
   mjGEOM_SKIN,                    // skin
   mjGEOM_LABEL,                   // text label
+  mjGEOM_TRIANGLE,                // triangle connecting a frame
 
   mjGEOM_NONE         = 1001      // missing geom type
 } mjtGeom;
@@ -742,6 +743,8 @@ struct mjModel_ {
   mjtNum*   cam_mat0;             // global orientation in qpos0              (ncam x 9)
   int*      cam_resolution;       // [width, height] in pixels                (ncam x 2)
   mjtNum*   cam_fovy;             // y-field of view (deg)                    (ncam x 1)
+  float*    cam_intrinsic;        // [focal length; principal point]          (ncam x 4)
+  float*    cam_sensorsize;       // sensor size                              (ncam x 2)
   mjtNum*   cam_ipd;              // inter-pupilary distance                  (ncam x 1)
   mjtNum*   cam_user;             // user data                                (ncam x nuser_cam)